+++ .github/workflows/pythonapp.yml
@@ -56,7 +56,7 @@ jobs:
       run: |
         python -m pip install torch==1.4 -f https://download.pytorch.org/whl/cpu/torch_stable.html
         # min. requirements for windows instances
-        python -c "f=open('requirements-dev.txt', 'r'); txt=f.readlines(); f.close(); print(txt); f=open('requirements-dev.txt', 'w'); f.writelines(txt[1:10]); f.close()"
+        python -c "f=open('requirements-dev.txt', 'r'); txt=f.readlines(); f.close(); print(txt); f=open('requirements-dev.txt', 'w'); f.writelines(txt[1:11]); f.close()"
     - name: Install the dependencies
       run: |
         python -m pip install torch==1.4
