+++ docs/source/installation.md
@@ -126,10 +126,9 @@ Since MONAI v0.2.0, the extras syntax such as `pip install 'monai[nibabel]'` is 
 
 - The options are
 ```
-[nibabel, skimage, pillow, tensorboard, ignite]
+[nibabel, skimage, pillow, tensorboard, gdown, ignite]
 ```
-which correspond to `nibabel`, `scikit-image`, `pillow`, `tensorboard`, and `pytorch-ignite` respectively.
+which correspond to `nibabel`, `scikit-image`, `pillow`, `tensorboard`, 
+`gdown`, and `pytorch-ignite` respectively.
 
 - `pip install 'monai[all]'` installs all the optional dependencies.
-
-
