+++ setup.cfg
@@ -22,6 +22,7 @@ all =
     pillow
     tensorboard
     pytorch-ignite==0.3.0
+    gdown==3.11.1
 nibabel =
     nibabel
 skimage =
@@ -30,6 +31,8 @@ pillow =
     pillow
 tensorboard =
     tensorboard
+gdown =
+    gdown==3.11.1
 ignite =
     pytorch-ignite==0.3.0
