+++ tests/test_check_md5.py
@@ -16,7 +16,7 @@ import numpy as np
 import tempfile
 from PIL import Image
 from parameterized import parameterized
-from monai.application import check_md5
+from monai.apps import check_md5
 
 TEST_CASE_1 = ["f38e9e043c8e902321e827b24ce2e5ec", True]
